take out cmds & filters, move out orderColList
